// src/types/enums.ts
export enum Status {
  SUNNY = "white-balance-sunny",
  RAINY = "weather-pouring",
  CLOUDY = "weather-cloudy",
  THUNDERSTORMS = "weather-lightning-rainy",
  FOGGY = "weather-fog",
  WINDY = "weather-windy",
  SNOWY = "weather-snowy",
  HAIL = "weather-hail",
  PARTLY_CLOUDY = "weather-partly-cloudy",
}

export enum Store {
  LOCATION = "LOCATION",
}